appengine/app/requirements.txt
webapp2==2.5.2
google-cloud-datastore==1.15.3
play-scraper==0.6.0
WebOb==1.8.6
futures==3.3.0
grpcio==1.33.1
requests-toolbelt==0.9.1
lxml==4.6.1